services/parser-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pageindex==0.2.5.1
aiokafka==0.10.0
boto3==1.34.0
openai>=1.70.0
anthropic>=0.40.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.0
pytest-asyncio==0.23.0
httpx==0.26.0
aiofiles==23.2.1
PyPDF2==3.0.1
